(#3925) - more refactoring
* use camelCase for inner modules, ala lodash
* rename stuff to `index.js` where it makes sense
