+++ lib/adapter.js
@@ -7,8 +7,8 @@ var EventEmitter = require('events').EventEmitter;
 var upsert = require('./deps/upsert');
 var Changes = require('./changes');
 var Promise = utils.Promise;
-var isDeleted = require('./deps/docs/is-deleted');
-var isLocalId = require('./deps/docs/is-local-id');
+var isDeleted = require('./deps/docs/isDeleted');
+var isLocalId = require('./deps/docs/isLocalId');
 
 /*
  * A generic pouch adapter
