+++ lib/adapters/http/index.js
