+++ lib/adapters/idb/allDocs.js
@@ -2,8 +2,8 @@
 
 var merge = require('../../merge');
 var errors = require('../../deps/errors');
-var idbUtils = require('./idb-utils');
-var idbConstants = require('./idb-constants');
+var idbUtils = require('./utils');
+var idbConstants = require('./constants');
 
 var ATTACH_STORE = idbConstants.ATTACH_STORE;
 var BY_SEQ_STORE = idbConstants.BY_SEQ_STORE;
