+++ lib/adapters/idb/blobSupport.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var utils = require('../../utils');
-var idbConstants = require('./idb-constants');
+var idbConstants = require('./constants');
 var DETECT_BLOB_SUPPORT_STORE = idbConstants.DETECT_BLOB_SUPPORT_STORE;
 
 //
