+++ lib/adapters/idb/bulkDocs.js
@@ -3,11 +3,11 @@
 var utils = require('../../utils');
 var errors = require('../../deps/errors');
 var preprocessAttachments =
-  require('../../deps/docs/preprocess-attachments');
-var processDocs = require('../../deps/docs/process-docs');
-var isLocalId = require('../../deps/docs/is-local-id');
-var idbUtils = require('./idb-utils');
-var idbConstants = require('./idb-constants');
+  require('../../deps/docs/preprocessAttachments');
+var processDocs = require('../../deps/docs/processDocs');
+var isLocalId = require('../../deps/docs/isLocalId');
+var idbUtils = require('./utils');
+var idbConstants = require('./constants');
 
 var ATTACH_AND_SEQ_STORE = idbConstants.ATTACH_AND_SEQ_STORE;
 var ATTACH_STORE = idbConstants.ATTACH_STORE;
