+++ lib/adapters/idb/constants.js
