+++ lib/adapters/idb/index.js
@@ -2,14 +2,14 @@
 
 var utils = require('../../utils');
 var merge = require('../../merge');
-var isDeleted = require('../../deps/docs/is-deleted');
-var isLocalId = require('../../deps/docs/is-local-id');
+var isDeleted = require('../../deps/docs/isDeleted');
+var isLocalId = require('../../deps/docs/isLocalId');
 var errors = require('../../deps/errors');
-var idbUtils = require('./idb-utils');
-var idbConstants = require('./idb-constants');
-var idbBulkDocs = require('./idb-bulk-docs');
-var idbAllDocs = require('./idb-all-docs');
-var checkBlobSupport = require('./idb-blob-support');
+var idbUtils = require('./utils');
+var idbConstants = require('./constants');
+var idbBulkDocs = require('./bulkDocs');
+var idbAllDocs = require('./allDocs');
+var checkBlobSupport = require('./blobSupport');
 
 var ADAPTER_VERSION = idbConstants.ADAPTER_VERSION;
 var ATTACH_AND_SEQ_STORE = idbConstants.ATTACH_AND_SEQ_STORE;
