+++ lib/adapters/idb/utils.js
@@ -4,7 +4,7 @@ var errors = require('../../deps/errors');
 var utils = require('../../utils');
 var base64 = require('../../deps/binary/base64');
 var btoa = base64.btoa;
-var constants = require('./idb-constants');
+var constants = require('./constants');
 var readAsBinaryString = require('../../deps/binary/readAsBinaryString');
 var b64StringToBlob = require('../../deps/binary/base64StringToBlobOrBuffer');
