+++ lib/adapters/leveldb/index.js
@@ -18,9 +18,9 @@ requireLeveldown();
 var errors = require('../../deps/errors');
 var merge = require('../../merge');
 var utils = require('../../utils');
-var isDeleted = require('../../deps/docs/is-deleted');
-var isLocalId = require('../../deps/docs/is-local-id');
-var processDocs = require('../../deps/docs/process-docs');
+var isDeleted = require('../../deps/docs/isDeleted');
+var isLocalId = require('../../deps/docs/isLocalId');
+var processDocs = require('../../deps/docs/processDocs');
 var md5 = require('../../deps/md5');
 var migrate = require('../../deps/migrate');
 var Deque = require("double-ended-queue");
@@ -29,7 +29,7 @@ var readAsBinaryString = require('../../deps/binary/readAsBinaryString');
 var binStringToBluffer =
   require('../../deps/binary/binaryStringToBlobOrBuffer');
 
-var LevelTransaction = require('./leveldb-transaction');
+var LevelTransaction = require('./transaction');
 
 var DOC_STORE = 'document-store';
 var BY_SEQ_STORE = 'by-sequence';
