+++ lib/adapters/leveldb/transaction.js
