+++ lib/adapters/websql/bulkDocs.js
@@ -3,12 +3,12 @@
 var utils = require('../../utils');
 var errors = require('../../deps/errors');
 var preprocessAttachments =
-  require('../../deps/docs/preprocess-attachments');
-var isLocalId = require('../../deps/docs/is-local-id');
-var processDocs = require('../../deps/docs/process-docs');
+  require('../../deps/docs/preprocessAttachments');
+var isLocalId = require('../../deps/docs/isLocalId');
+var processDocs = require('../../deps/docs/processDocs');
 
-var websqlUtils = require('./websql-utils');
-var websqlConstants = require('./websql-constants');
+var websqlUtils = require('./utils');
+var websqlConstants = require('./constants');
 
 var DOC_STORE = websqlConstants.DOC_STORE;
 var BY_SEQ_STORE = websqlConstants.BY_SEQ_STORE;
