+++ lib/adapters/websql/constants.js
