+++ lib/adapters/websql/index.js
@@ -2,15 +2,15 @@
 
 var utils = require('../../utils');
 var merge = require('../../merge');
-var isDeleted = require('../../deps/docs/is-deleted');
-var isLocalId = require('../../deps/docs/is-local-id');
+var isDeleted = require('../../deps/docs/isDeleted');
+var isLocalId = require('../../deps/docs/isLocalId');
 var errors = require('../../deps/errors');
-var parseHexString = require('../../deps/parse-hex');
+var parseHexString = require('../../deps/parseHex');
 var binStringToBlob = require('../../deps/binary/binaryStringToBlobOrBuffer');
 
-var websqlConstants = require('./websql-constants');
-var websqlUtils = require('./websql-utils');
-var websqlBulkDocs = require('./websql-bulk-docs');
+var websqlConstants = require('./constants');
+var websqlUtils = require('./utils');
+var websqlBulkDocs = require('./bulkDocs');
 
 var ADAPTER_VERSION = websqlConstants.ADAPTER_VERSION;
 var DOC_STORE = websqlConstants.DOC_STORE;
