+++ lib/adapters/websql/utils.js
@@ -3,7 +3,7 @@
 var utils = require('../../utils');
 var errors = require('../../deps/errors');
 
-var websqlConstants = require('./websql-constants');
+var websqlConstants = require('./constants');
 
 var BY_SEQ_STORE = websqlConstants.BY_SEQ_STORE;
 var ATTACH_STORE = websqlConstants.ATTACH_STORE;
