+++ lib/changes.js
@@ -1,6 +1,6 @@
 'use strict';
 var utils = require('./utils');
-var isDeleted = require('./deps/docs/is-deleted');
+var isDeleted = require('./deps/docs/isDeleted');
 var merge = require('./merge');
 var errors = require('./deps/errors');
 var EE = require('events').EventEmitter;
