+++ lib/deps/docs/isLocalId.js
