+++ lib/deps/docs/preprocessAttachments.js
