+++ lib/deps/docs/processDocs.js
@@ -2,9 +2,9 @@
 
 var merge = require('../../merge');
 var errors = require('../errors');
-var updateDoc = require('./update-doc');
-var isDeleted = require('./is-deleted');
-var isLocalId = require('./is-local-id');
+var updateDoc = require('./updateDoc');
+var isDeleted = require('./isDeleted');
+var isLocalId = require('./isLocalId');
 
 var collections = require('pouchdb-collections');
 var Map = collections.Map;
