+++ lib/deps/docs/updateDoc.js
@@ -2,8 +2,8 @@
 
 var merge = require('../../merge');
 var errors = require('../errors');
-var isDeleted = require('./is-deleted');
-var parseDoc = require('./parse-doc').parseDoc;
+var isDeleted = require('./isDeleted');
+var parseDoc = require('./parseDoc').parseDoc;
 
 function revExists(metadata, rev) {
   var found = false;
