+++ lib/deps/md5.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var toPromise = require('./to-promise');
+var toPromise = require('./toPromise');
 var base64 = require('./binary/base64');
 var crypto = require('crypto');
 var Md5 = require('spark-md5');
