+++ lib/deps/migrate.js
@@ -2,7 +2,7 @@
 
 var fs = require('fs');
 var path = require('path');
-var isLocalId = require('./docs/is-local-id');
+var isLocalId = require('./docs/isLocalId');
 var merge = require('../merge');
 var levelup = require('levelup');
 var through = require('through2').obj;
