+++ lib/deps/parseHex.js
