+++ lib/deps/toPromise.js
