+++ lib/index.js
@@ -10,15 +10,15 @@ PouchDB.Errors = require('./deps/errors');
 PouchDB.replicate = require('./replicate').replicate;
 PouchDB.sync = require('./sync');
 PouchDB.version = require('./version');
-var httpAdapter = require('./adapters/http/http');
+var httpAdapter = require('./adapters/http');
 PouchDB.adapter('http', httpAdapter);
 PouchDB.adapter('https', httpAdapter);
 
-PouchDB.adapter('idb', require('./adapters/idb/idb'), true);
-PouchDB.adapter('websql', require('./adapters/websql/websql'), true);
+PouchDB.adapter('idb', require('./adapters/idb'), true);
+PouchDB.adapter('websql', require('./adapters/websql'), true);
 PouchDB.plugin(require('pouchdb-mapreduce'));
 
 if (!process.browser) {
-  var ldbAdapter = require('./adapters/leveldb/leveldb');
+  var ldbAdapter = require('./adapters/leveldb');
   PouchDB.adapter('leveldb', ldbAdapter, true);
 }
