+++ lib/plugins/levelalt.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var LevelPouch = require('../adapters/leveldb/leveldb');
+var LevelPouch = require('../adapters/leveldb/index');
 var utils = require('../utils');
 module.exports = altFactory;
 function altFactory(adapterConfig, downAdapter) {
