+++ lib/replicate/replicate.js
@@ -3,8 +3,8 @@
 var utils = require('./../utils');
 var Checkpointer = require('./checkpointer');
 var backOff = require('./backoff');
-var genReplicationId = require('./gen-replication-id');
-var getDocs = require('./get-docs');
+var genReplicationId = require('./genReplicationId');
+var getDocs = require('./getDocs');
 
 function replicate(src, target, opts, returnValue, result) {
   var batches = [];               // list of batches to be processed
