+++ lib/utils.js
@@ -10,7 +10,7 @@ var EventEmitter = require('events').EventEmitter;
 var collections = require('pouchdb-collections');
 exports.Map = collections.Map;
 exports.Set = collections.Set;
-var parseDoc = require('./deps/docs/parse-doc');
+var parseDoc = require('./deps/docs/parseDoc');
 
 var Promise = require('./deps/promise');
 exports.Promise = Promise;
@@ -224,7 +224,7 @@ Changes.prototype.notify = function (dbName) {
 
 exports.once = require('./deps/once');
 
-exports.toPromise = require('./deps/to-promise');
+exports.toPromise = require('./deps/toPromise');
 
 exports.adapterFun = function (name, callback) {
   var log = require('debug')('pouchdb:api');
@@ -362,7 +362,7 @@ exports.info = function (str) {
   }
 };
 
-exports.parseUri = require('./deps/parse-uri');
+exports.parseUri = require('./deps/parseUri');
 
 exports.compare = function (left, right) {
   return left < right ? -1 : left > right ? 1 : 0;
